Added teleport effects to level teleports.

+++ include/you.h
@@ -332,7 +332,7 @@ struct you {
     xchar ux0, uy0;     /* initial position FF */
     d_level uz, uz0;    /* your level on this and the previous turn */
     d_level utolev;     /* level monster teleported you to, or uz */
-    uchar utotype;      /* bitmask of goto_level() flags for utolev */
+    short utotype;      /* bitmask of goto_level() flags for utolev */
     boolean umoved;     /* changed map location (post-move) */
     int last_str_turn;  /* 0: none, 1: half turn, 2: full turn
                            +: turn right, -: turn left */
